Migrated rtos demo code to Archive/ to prevent merge conflicts
